requirements.txt
amqp==2.6.0
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.6
Django==2.2.13
django-environ==0.4.5
django-filter==2.0.0
django-storages==1.9.1
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
kombu==4.6.11
Pillow==6.2.1
psycopg2-binary==2.8.4
pytz==2019.3
redis==3.5.3
sqlparse==0.3.0
vine==1.3.0
whitenoise==4.1.4
